image-classification/requirements.txt
einops==0.7.0
loralib==0.1.1
matplotlib==3.8.0
numpy==1.23.5
opencv_python==4.8.0.76
opencv_python_headless==4.7.0.72
pandas==1.2.4
Pillow==9.3.0
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
timm==0.9.2
torch==2.0.0
torchvision==0.15.1
vit_pytorch==1.2.4
wandb
Ipython
